consumer/requirements.txt
opentelemetry-distro[otlp]==0.28b1
opentelemetry-instrumentation-kafka-python==0.28b1
opentelemetry-instrumentation-logging==0.28b1
kafka-python==2.0.2